Task
The program allows you to create a complex shape that consists of a pentagon and a circle inscribed in it. The project implements drawing functions with windows.h and windowsx.h, moving and hiding functions for each figure. A container class (stack as a list) for the shapes is also created.